Separate the media-specific methods out from ObjectTypeParsingRules
Keep OTPR focused on the value class functionality. Move the utility
methods for construction from various data sources to a utility class.
